Script Editor: add another TODO
Now that we have such a nice editor, we should make full use of it.  So
let's integrate it much better with Fiji Build, so that it detects
automatically when it is in src-plugins/ in a fiji.git checkout, and
run Fiji Build with debug=true.  That way, we should be able to use
the debugger without further problems!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>