Script Editor: when re-opening the Find dialog, select the find pattern
When the Find dialog is re-opened, usually you either want to search the
same thing as last time (in which case the patterns hould be retained), or
you search for something completely different, in which case it is very
convenient if the pattern is already selected so you can type the pattern
right away.

Make it so.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>