Fakefile: use simple one-size-fits-all rule for non-submodule targets
Earlier, we had custom rules for all plugins, mainly because we wanted
to calculate the strip path (i.e. what prefix to strip when putting
the files into the .jar) from the Java classes (determining the prefix
using the Java package). So the Java files _had_ to come first.

Now that Fiji Build is less magic, but looks at the first "**" in the
prerequisites to determine the prefix, we need no longer have those
rules and can have two simple rules for .jar files in plugins/ and jars/.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>