Script Editor: allow developing the script editor itself
There was a cute problem with the RSyntaxTextArea's action map being
cached, even when the new RSyntaxTextArea was loaded from a new class
loader, and therefore the 'instanceof RTextArea' check within the action
failed. Consequently, you could not edit anything; the keyboard input
was just discarded.

Now, just before forcing a new class loader, we force the action map
back to uncached (so that the formerly installed map is still
operational).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>