Script Editor: lazily load tools.jar via internet when asked for it
Usually you do not want to debug Java code when running the Script
Editor. Therefore, it is pointless to load tools.jar via internet,
especially if the connection to the Fiji server is slow.

The obvious fix: do not do it automatically, but only when asked for.

Noticed by Yıldırım Karslıoğlu.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>