Script Editor: rip out a lot of inefficient completion code
At the moment, the auto-completion code is fantastically non-functional.
The only use we have for it is to get at the full class names given
their base name, for the "Open Help for Class" and "Open .java for
Class" functions.

So let's concentrate on that, and avoid all expensive operations during
class discovery (such as splitting each and every class name into
individual parts by a regular expression).

While at it, the class ClassNames now no longer pretends that Java
packages are hierarchical, and is starting to be the simple map from
package names to simple lists of class names it should have been from
the start.

This is just the very first step in a probably long but necessary
cleanup of the completion code.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>