Fiji Updater: use ReadyForUpload directly
As before, to avoid dependencies for the Fiji Updater on Fiji-specific
stuff including Beanshell scripts, the Show Changes functionality of the
Fiji Updater lives in the Script Editor plugin. However, this allows us to
avoid running the test whether a file is ready to be uploaded in a
Beanshell interpreter; we can use the ReadyForUpload class which was just
added for that purpose.

Incidentally, this makes the check much, much quicker.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>