Fiji Updater "Rebuild": put the text into a scroll panel
They output of a rebuild can get quite long and wide, especially now that
we build MacAdapter in ImageJ which spits out 10 verbose warnings about an
ambiguity using varargs. When such a long output was displayed, the window
tended to get very "tall" and the "Okay" button vanished below the bottom
of the screen.

While at it, do not repeat that frame.getContentPane() call several times.

Also, make sure that the button is centered...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>