Script editor: allow customization of the git path
This can be done via the "git.path" environment variable.

So for example, you can set it using a Beanshell script in the Script
Editor with the following one-liner:

	System.setProperty(fiji.scripting.FileFunctions.GIT_PATH_PROP,
		"/path/to/my/git");

Afterwards, for the duration of the Fiji session, the Git menu in the
Script Editor should work.

Note that for the commands we currently support (all Git built-ins),
there is no problem with calling git via an absolute path like this.
However, some Git commands leverage shell scripts, Python scripts, etc.,
internally, and will not work properly without other stuff in the path.
(This would only be a problem if additional non-built-in git commands
were to be added to the Git menu, or in certain isoteric situations such
as trying to use Git connected to a remote Mercurial repository.)