Script Editor: remove another Fiji'ism
The Script Editor used to be the only way to develop Fiji itself (before
we switched to Maven project configurations). That included the
development of the Script Editor itself, which required some hackery to
avoid menu accelerator issues when there were two editors open, loaded
in different class loaders, but registered in Swing by the same name.

An unfortunate side effect was that under certain circumstances that
were never properly described, the accelerators stopped working
altogether.

Happily, the primary way to develop ImageJ2 is not in the Script Editor,
but in any Maven-enabled IDE, so we can get rid of the hack and make the
editor more robust again in the same run.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>