Fix inconsistent typing of PluginInfos
They were supposed to always be typed with a "PT extends ImageJPlugin"
that is the plugin *type* (e.g., Command or Display), but unfortunately
in some places it was conflated with a "P extends ImageJPlugin" with P
being used the actual plugin *class* (e.g., AboutImageJ). There were
methods of PluginService that took a "Class<P> pluginClass" but returned
a "PluginInfo<P>" which is bogus.

Now, whenever a plugin *type* is required, the variable PT is used, and
whenever a typed plugin *class* is required, the variable P is used.