Explicitly declare plugin type for Commands
This will be required in the future, both to avoid confusion, and
because the SciJava plugin framework is more general than just ImageJ.

It will now be less likely for someone to accidentally declare a plugin
of an unintended type. Unfortunately, it is slightly less simple (since
"@Plugin" with no arguments is no longer legal), but I think the
tradeoff is worthwhile.

This change also cleans up imports and header whitespace in a few
places.