Upgrade to commons-math 3.2
This version is available from Maven Central as commons-math3 with a
different package prefix: org.apache.commons.math3. This has the
advantage of avoiding conflict with Apache Commons Math 2.x, which is
also available as artifact commons-math with package prefix
org.apache.commons.math.

The three plugins that depend on commons-math (Script Editor, Simple
Neurite Tracer and Trainable Segmentation) were updated to use the new
package prefix and API where applicable.

Right now, there is nothing in Fiji core that uses commons-math 2.x,
but thanks to the package shading, it is now possible for third party
plugins to depend on 2.x without interfering with the ones using 3.x.