Stop using deprecated CommandService#run methods
Many of these changes (the ones labeled FIXME) will need to be tweaked
again, after the deprecated run API has been removed from ModuleService
and CommandService. The reason is that as things stand, there is an
ambiguity between the methods with signature (X, boolean, Object...) and
the deprecated ones with corresponding (X, Object...) with no boolean.
Unfortunately, due to autoboxing, the Java compiler has no way of
knowing if a passed boolean was intended to fill the actual boolean
parameter, or instead become part of the Object varargs array.

From my naive perspective, I think it should prefer to fill the boolean
parameter, the same as would happen with two non-varargs method
signatures that take a boolean and an Object respectively. But it
doesn't, and the problem will go away once the deprecated API is
removed, so we can adjust the method calls again after that point.