TextEditor: guard against language NPEs
When the Script Editor initially comes up, the ScriptLanguage is null.
So we need to be careful when checking whether the current language is
a compiled language. It is also possible for it to get set back to null
(i.e., "None") by the user, which must also not throw an NPE.