Port findResources() method from deprecated class
ImageJ's AppUtils class has been deprecated in favor of SciJava's, but
the findResources() method has not made it there yet.

It is actually not quite well-exercised enough yet to graduate to
SciJava common; Besides, it is not at all clear what class would be
its appropriate destination (AppUtils? FileUtils? ResourceUtils?).

So let's develop this method in imagej-ui-swing until it is stable, and
move it then.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>