Bump to pom-imagej 2.15 to get a newer version of scripting-java
This is needed to support developing .java plugins in the script
editor, as scripting-java and MiniMaven got bug fixes for that use
case in the meantime. The updated versions are already referenced
in a new pom-scijava, but pom-imagej was not yet updated accordingly
(waiting for the very topic branches in imagej-ui-swing and
imagej-legacy for which the bug fixes were developed).

The pom-imagej version 2.15 already referenced the imagej-ui-swing
version we are about to release; it has been bumping the versions
preemptively.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>