EditorPane: create addHeader flag
We do not always want to create a header - for example, when opening a
template or a file from disk.

The ONLY situation when we want to add headers is when going from a new
file (language = null) to a manually selected language.

Adding an addHeader flag allows us to ensure headers are not added by
mistake.