InterpreterWindow: use context injection
Rather than passing a hardcoded list of services, let's declare them
using @Parameter. That way, when we inevitably need new services in the
future, we do not have to change the API contract in order to pass them
to the constructor, but can instead declare them as new instance fields.

What new services will we need? Well, the PrefService for starters,
since we'll need to persist the interpreter history and active tab
settings.