Don't run unsaved compiled language
Compiled scripts (e.g. java) need to be saved before they can be run, so
the compiler can operate on the code. Thus the handleUnsavedChanges
method should return different values if it's being called before
compilation (false) or before quit (true is OK because the user
shouldn't be forced to save before exit).

An appropriate warning is then printed if the file is not saved