TextEditor: avoid < 0 indicies
If no tab is currently selected (or exists) in the ScriptEditor then
getTab will see a negative selectedIndex.

In this case, in addition to creating and selecting a new tab, we must
update the index variable as the negative value is no longer valid.

Fixes http://fiji.sc/bugzilla/show_bug.cgi?id=1174