Create OpViewer
Adds an OpViewer, a tree-based view of all available ops. This list is
sorted by namespace and op type, with each op signature as leaves in the
tree.

The OpViewer is exposed via a companion BrowseOps command.

The goal of this addition is to provide users an easy method to find out
what ops are available, while creatign possibilities for future
functionality - e.g. via selection listeners to do code insertion, or
open javadoc.