Organize Templates menu by topic and language
This adds a "[by language]" submenu with all the same templates, in a
hierarchy based on the script language first and foremost.

It keeps the original layout as well, appending the language as a
parenthetical note for each script.

In this way, we allow browsing of the Templates menu by category and
script name, without needing to choose a single hierarchy for all cases.