Hide REPL pane unless "persistent" is checked
This fixes a bug where resizing the tab's JSplitPane divider to make the
output pane taller instead results in the REPL "emerging" while in
non-persistent mode. This was happening because the screenAndPromptSplit
prioritized keeping its JSplitPane divider at the same integer value,
resulting in the output pane height staying constant and the REPL pane's
height increasing. By setting the REPL pane to invisible, the resizing
logic now instead gives the extra space to the output pane, as desired.