Improve support for Java 9+
This avoids the hardcoded message to stderr:

  [ERROR]: Cannot locate JRE jar in /Library/Java/JavaVirtualMachines/jdk-10.0.1.jdk/Contents/Home

And subsequent exception:

  java.lang.IllegalArgumentException: info cannot be null
    at org.fife.rsta.ac.java.JarManager.addClassFileSource(JarManager.java:157)
    at org.fife.rsta.ac.java.JarManager.addCurrentJreClassFileSource(JarManager.java:193)
    at org.scijava.ui.swing.script.languagesupport.JavaLanguageSupportPlugin.<init>(JavaLanguageSupportPlugin.java:56)

Which happens because the LibraryInfo.getJreJarInfo(File) method fails
to locate the main JRE JAR file, because Java 9+ does not have one --
rather, core classes are divided into modules in the jmods folder.

And allows the plugin to be created successfully anyway.

Fixes #29.

It also makes a best effort to find the "main JRE JAR" -- which for the
moment we point at jmods/java.base.jmod, one of many core modules --
and the sources archive in its new location, which is lib/src.zip.

Unfortunately, this logic is still insufficient to actually allow
RSyntaxTextArea's language support to display javadoc for core Java
classes with Java 9+, because the src.zip's internal directory structure
is now divided by module, which the library does not yet understand.

So really, an upstream fix will be needed to address this. Someday.