Create new menu item "Source or javadoc for class or package..." which fixes or replaces the "Open Help" menu item, by using ClassUtils and showing a UI that lists all matching classes with buttons to open their source code in github or gitlab and the javadoc when available (will guess the URL, load it, and not show a link to it when 404).
Can match partial text (the list of possible classes may be longer)
and can also match package names or partial package names. Any substring
of a fully qualified class name will be matched, which is fabulous.

(Compare with current implementatio nof "Open Help..." which requires
a fully qualified class name to work, which is not useful as often
one does not know it or not have it handy in a script.)