Implement "Mark occurrences" by double-click and Error Strip
The Mark All' occurrences feature is restricted to certain identifiers defined on individually for each language.
This makes it so, that in addition to whatever is being specified, it is also possible to double-click on any string
to highlight all occurrences. In addition, this adds an error strip to the right side of the text area, marking the positioning of lines associated with syntax errors and 'Mark all'/'Find all' occurrences.

Related to @acardona's report, https://github.com/scijava/script-editor/pull/56#issuecomment-1050888837