Extend error tracking to more languages and make console searchable
This should cover all but IJM (which anyway is not really supported) and 'R'.
Awkwardly, traces from the renjin engine contain little detail so there is not much we can do there.

This also adds commands to highlight, delete, and search contents of the output/error panel.
While at it,  make some effort to make highlighted colors work with any theme