Fix ImageJ macro language functionality
The language rename from "IJ1 Macro" to "ImageJ Macro"
(imagej/imagej-legacy@a03160552a945044a0c472ccde27d54bd76432af) broke
the Script Editor's special handling, because internally, the name was
being used to check whether the language is the ImageJ macro language.

It's unfortunate that this component has any special handling for the
Image Macro language whatsoever -- it's supposed to be completely
agnostic to specific script languages -- but we are where we are, and
this commit fixes the immediate problem.

See also: https://forum.image.sc/t/71733