Update to RSyntaxTextArea v3.6.0
This fixes the issue detailed in
https://github.com/scijava/pom-scijava/pull/288.

Context:
In newer versions of RSyntaxtTextArea, AbstractJFlexCTokenMaker adds an
abstract yystate() method intended to return the current state of a
JFlex lexer. Since it is abstract, it must be implemented by any
concrete subclass, so we patch ImageJMacroTokenMaker so that the current
zzLexicalState is returned.

Tested with: `mvn clean package -Denforcer.skip`

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>